+++ tests/exc_utils.py
@@ -20,6 +20,17 @@ def get_exception_info_2(*args):
         return ExceptionInfo.new()
 
 
+def get_exception_info_with_locals(*args):
+    local_1 = args[0]
+
+    try:
+        raise Exception(*args)
+        local_undefined = args[2]
+    except:
+        local_except = args[1]
+        return ExceptionInfo.new()
+
+
 def get_exception_info_with_2_frames(c_var=None, z_var=None, a_var=None, *extras, **named):
     try:
         _nested_function(c_var, z_var, a_var, **named)