+++ tests/test_frame.py
@@ -6,7 +6,8 @@ from __future__ import division, print_function, unicode_literals
 
 import unittest
 
-from exc_utils import get_exception_info_1, get_exception_info_with_2_frames
+from exc_utils import (get_exception_info_1, get_exception_info_with_2_frames,
+    get_exception_info_with_locals)
 from diagnostics._py3k import unicode
 
 
@@ -34,3 +35,35 @@ class TestFrame(unittest.TestCase):
         expected = ("b_var", "c_var", "a_var", "**named")
         self.assertEqual(arguments, expected)
 
+    def test_frame_properties(self):
+        info = get_exception_info_with_locals(1, 2, 3)
+
+        frame = info.frames[0]
+        self.assertTrue(isinstance(frame.number, int))
+        self.assertEqual(frame.number, 1)
+        self.assertEqual(repr(frame), "<Frame#1: get_exception_info_with_locals>")
+
+    def test_frame_locals(self):
+        info = get_exception_info_with_locals(22.22, 33.5)
+
+        frame = info.frames[0]
+        self.assertEqual(len(frame.locals), 2)
+
+        self.assertEqual(frame.locals[0].name, "local_1")
+        self.assertEqual(frame.locals[0].value, 22.22)
+
+        self.assertEqual(frame.locals[1].name, "local_except")
+        self.assertEqual(frame.locals[1].value, 33.5)
+
+    def test_frame_globals(self):
+        info = get_exception_info_with_locals(22.22, 33.5)
+
+        frame = info.frames[0]
+        self.assertEqual(len(frame.globals), 6)
+
+        self.assertEqual(frame.globals[0].name, "absolute_import")
+        self.assertEqual(frame.globals[1].name, "division")
+        self.assertEqual(frame.globals[2].name, "EXCEPTION_FILE_NAME_1")
+        self.assertEqual(frame.globals[3].name, "EXCEPTION_FILE_NAME_2")
+        self.assertEqual(frame.globals[4].name, "print_function")
+        self.assertEqual(frame.globals[5].name, "unicode_literals")